exercice 21

+++ controllers/wood.js
@@ -32,7 +32,6 @@ exports.addWood = async (req, res) => {
 
 exports.updateWood = async (req, res) => {
     try {
-        const pathname = `${req.protocol}://${req.get("host")}/uploads/${req.file.filename}`;
         const woodId = req.params.id;
         const updatedWoodData = JSON.parse(req.body.datas);
         const woodToUpdate = await Wood.findByPk(woodId);
@@ -42,19 +41,50 @@ exports.updateWood = async (req, res) => {
         }
 
         if (req.file) {
-            const oldImagePath = woodToUpdate.image.split('/uploads/')[1];
-            fs.unlink(`./uploads/${oldImagePath}`, (err) => {
-                if (err) {
-                    return res.status(401).json({ message: 'Erreur lors de la suppression de l\'ancienne image :', err });
-                }
-            });
+            const pathname = `${req.protocol}://${req.get("host")}/uploads/${req.file.filename}`;
+            if (woodToUpdate.image) {
+                const oldImagePath = woodToUpdate.image.split('/uploads/')[1];
+                fs.unlink(`./uploads/${oldImagePath}`, (err) => {
+                    if (err) {
+                        return res.status(401).json({ message: 'Erreur lors de la suppression de l\'ancienne image :', err });
+                    }
+                });
+            }
             updatedWoodData.image = pathname;
-            const updatedWood = await woodToUpdate.update(updatedWoodData);
-            res.status(200).json(updatedWood);
         }
+
+        const updatedWood = await woodToUpdate.update(updatedWoodData);
+        res.status(200).json(updatedWood);
 
     } catch (error) {
         console.error(error);
         res.status(500).json({ message: 'Une erreur est survenue lors de la mise à jour de l\'essence de bois.' });
+    }
+};
+
+exports.deleteWood = async (req, res) => {
+    try {
+        const woodId = req.params.id;
+        const woodToDelete = await Wood.findByPk(woodId);
+
+        if (!woodToDelete) {
+            return res.status(404).json({ message: 'Essence de bois non trouvée.' });
+        }
+
+        if (woodToDelete.image) {
+            const oldImagePath = woodToDelete.image.split('/uploads/')[1];
+            fs.unlink(`./uploads/${oldImagePath}`, (err) => {
+                if (err) {
+                    return res.status(401).json({ message: 'Erreur lors de la suppression de l\'image associée :', err });
+                }
+            });
+        }
+
+        await woodToDelete.destroy();
+        res.status(200).json({ message: 'Essence de bois supprimée avec succès.' });
+
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ message: 'Une erreur est survenue lors de la suppression de l\'essence de bois.' });
     }
 };
